Create README.md
Description :   A hero-image / banner meant for websites or mobile phones.
The sketch has voice activation and response to commands.  Speak the commands on the bottom of the screen, shake your phone or dragging/swiping  to advance the sketch to the next stage.  The final stage is an image/gif that is sharable on social media to promote the album.   The 4 stages are 
0. Asleep  -  Say WAKE UP
1. Awake  -  Say LISTEN
2. Listening - Say SHARE
3. Sharing. The end of the sketch.

During any stage press a mouse button or keyboard to access print media promotional material.

Note : Actual social media sharing not implemented and additional animations happen when the sketch detechs any sound from the mind.
Additional Note : Stars collect if alt-tabbed, and shoot out at once but this corrects over time.